requirements.txt
fastapi
httpx
python-dotenv
pillow
uvicorn[standard]
